pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "qdrant-av-edgecase-memory"
version = "0.1.0"
description = "Multi-modal edge-case retrieval for autonomous vehicles using Qdrant named vectors"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "qdrant-client>=1.7.0",
  "numpy>=1.24",
  "Pillow>=10.0",
  "scikit-learn>=1.3"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
